+++ NEWS.md
@@ -7,6 +7,8 @@ colorize the (heuristically found) strongest interaction.
 
 * 06/02/2021 Added function `shap.importance` to return mean absolute SHAP values per variable.
 
+* 10/02/2021 Added `jitter_width`, `jitter_height` and `alpha` to `shap.plot.dependence`.
+
 # SHAPforxgboost 0.1.0
 * 12/12/2020 Added support for LightGBM.